优化补拍界面并修复显示的BUG 全景模式(专业模式)增加出厂默认参数 此版本可以生产使用了。

+++ LG-06 (Motor)/APP/Src/User/GPIO.c
@@ -515,7 +515,7 @@ void shutter_processing(uint8_t satuss)
 
 void BLE_PWR_TurnOn(void)
 {
-	uint8_t tims=0;
+	uint16_t tims=0;
 	while(!BLE_STATU)
 	{
 		BLE_PWR_DOWN;
